Deletando voto de maneira dinamica
