+++ semana12/labeddit/src/components/button/dislike.js
@@ -1,20 +1,28 @@
 import axios from 'axios'
-import React from 'react'
+import React, { useContext } from 'react'
 import { BASE_URL, headers } from '../../constants/url'
+import {GlobalStateContext} from "../../globalstate/GlobalStateContext"
 
 function Dislike(props) {
+    const {DeleteVote,setVoted} = useContext(GlobalStateContext)
 
     const getDislike = (id) => {
-        console.log(headers)
-        const body = {direction: -1}
-        axios
-        .put(`${BASE_URL}/posts/${id}/votes`,body,headers)
-        .then((response) => {
-            console.log(response)
-        })
-        .catch((err) => {
-            alert(err.message)
-        })
+        
+        if(props.voted < 0 ){
+            DeleteVote(id)
+        }else{
+            const body = {direction: -1}
+            axios
+            .put(`${BASE_URL}/posts/${id}/votes`,body,headers)
+            .then((response) => {
+                setVoted(true)
+            })
+            .catch((err) => {
+                alert(err.message)
+            })
+        }
+
+
     }
 
     return (
